requirements.txt
pandas~=1.1.1
Pillow==7.2.0
boto3==1.16.10
botocore==1.19.10
certifi==2020.6.20
scikit-build==0.11.1
PyMySQL==0.10.1
pytesseract==0.3.6
PyVirtualDisplay==2.2
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
webdriver-manager==3.2.2
opencv-python==4.4.0.46
SpeechRecognition
redis
azcaptchaapi
timeout-decorator
google-cloud-vision==0.35.2
